feat(async): Take ownership of Vec instead of slice
Instead of `&'d mut [u8]`, take ownership over `Vec<u8>`. This solves a
lot of scoping issues. Take the following example:
There is a struct `Control` with an `AsyncGroup<'a>` field, having the livetime `'a`
of `Context`. If I now want to add a new transfer to this `AsyncGroup`
in a method `Control::add_control`, I create a `Vec<u8>` in this method.
Giving it's slice to a new `Transfer`, which I then add to `AsyncGroup`,
the vec must outlive Context, so it must already be defined before I
even get the Context i.e. before I really start doing anything.