feat(async): Add control transfer
Add Transfer::control to create and submit control transfers.
In addition to the other method's parameters, it takes control setup
parameters to build the setup packet (first 8 bytes of the data buffer).
Therefore instead of `buffer: &mut [u8]` it takes `vec: &mut Vec<u8>`,
as we need to be able to prepend 8 bytes to it.