fix(async): Handle OUT-Control-Packets in Transfer::actual()
Control packets with an out-direction can have a data blob. The buffer
then has 8 bytes of control-setup-header following the data blob.
`actual_length` as returned by libusb's async methods seems to be the
same as `wLength` specified in the control setup header.

Therefore if the packet is a control packet with it's direction being
`OUT`, `Transfer::actual()` checks if both lengths are the same and
returns an empty slice pointing to the end of the sent data blob.